Reverse - Even Rows and Odd Columns
The program must accept an integer matrix of size RxC as the input. The program must reverse the elements in the even rows in the matrix. Then the program must reverse the elements in the odd columns in the matrix. Finally, the program must print the modified matrix as the output.